web/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
pymongo==4.8.0
pandas==2.2.3
numpy==1.26.4
